Move archiver from framework to website
